FastAPI_Backend/requirements.txt
sqlalchemy
fastapi
uvicorn
fastapi-jwt-auth-v2
pydantic
google-genai
passlib
python-dotenv
bcrypt